add new test file for openshift metrics
